// src/quotes/enums/quote-documentation.enum.ts
export enum QuoteAttachmentType {
  IDENTIFICATION_DOCUMENT = 'identification_document',
  PROOF_OF_RESIDENCE = 'proof_of_residence',
  ACTIVITY_PHOTO = 'activity_photo',
  PROOF_OF_INCOME = 'proof_of_income',
}

export enum IncomeProofType {
  BANK_STATEMENT = 'bank_statement',
  PAYSLIP = 'payslip',
  INSS_BENEFIT = 'inss_benefit',
  MEI_DAS = 'mei_das',
}